+++ lib/commons/lifecycle.dart
@@ -1,11 +1,11 @@
 import 'dart:async';
 
-import 'fp.dart';
-
+/// Life Cycle Observer
 abstract class LifeCycleObserver {
   void observeLiveData<T>(LiveData<T> lv);
 }
 
+/// Live Data Structure
 class LiveData<T> {
   final String name;
   final T initialValue;
@@ -13,13 +13,15 @@ class LiveData<T> {
   Stream<T> _stream;
   T _currentValue;
 
-  String get _tag => '{{LiveData:$name}}';
-
-  LiveData(LifeCycleObserver lc, {T initValue, this.name})
+  LiveData(LifeCycleObserver lc, {T initValue, this.name, bool broadcast = true})
       : this.initialValue = initValue,
         this._currentValue = initValue {
     lc?.observeLiveData(this);
-    streamController = StreamController<T>.broadcast();
+    if (broadcast) {
+      streamController = StreamController<T>.broadcast();
+    } else {
+      streamController = StreamController<T>();
+    }
   }
 
   LiveData.fromStream(Stream stream, {T initValue, this.name})
@@ -45,76 +47,17 @@ class LiveData<T> {
     streamController.close();
   }
 
-  ToggleableSubscriber<T> subscribe(
+  StreamSubscription<T> subscribe(
     void onData(T event), {
     Function onError,
     void onDone(),
     bool cancelOnError,
   }) {
-    return ToggleableSubscriber<T>.start(
-      streamController.stream,
+    return streamController.stream.listen(
       onData,
       onError: onError,
       onDone: onDone,
       cancelOnError: cancelOnError,
     );
-  }
-
-  // Helper
-
-  LiveData<S> map<S>(S convert(T event), {S initialValue}) =>
-      LiveData.fromStream(stream.map(convert), name: name, initValue: initialValue);
-
-  static LiveData<T> merge<T>(
-    Map<Symbol, Stream<T>> streamsMap, {
-    String name,
-    T initialValue,
-  }) {
-    return LiveData.fromStream(mergeSymbol(streamsMap), name: name, initValue: initialValue);
-  }
-
-  static LiveData<T> zipMemorize<T>(
-    Map<Symbol, Stream<T>> streamsMap, {
-    String name,
-    T initialValue,
-  }) {
-    return LiveData.fromStream(memorizeBarrier(mergeSymbol(streamsMap)),
-        name: name, initValue: initialValue);
-  }
-}
-
-class ToggleableSubscriber<T> {
-  bool enable = true;
-
-  ToggleableSubscriber.start(
-    Stream<T> stream,
-    void onData(T event), {
-    Function onError,
-    void onDone(),
-    bool cancelOnError,
-  }) {
-    try {
-      stream.listen(
-        (T event) {
-          //print('enable $enable');
-          if (!enable) return;
-          onData(event);
-        },
-        onError: onError,
-        onDone: onDone,
-        cancelOnError: cancelOnError,
-      );
-    } catch (e) {
-    }
-  }
-
-  ToggleableSubscriber pause() {
-    enable = false;
-    return this;
-  }
-
-  ToggleableSubscriber resume() {
-    enable = true;
-    return this;
   }
 }