+++ test/mvvm_arch_test.dart
@@ -3,6 +3,5 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:mvvm_arch/mvvm_arch.dart';
 
 void main() {
-  test('no', () {
-  });
+  test('no', () {});
 }