+++ CHANGELOG.md
@@ -1,3 +1,17 @@
-## [0.0.1] - TODO: Add release date.
+#
+## [1.0.2] - 2020-04-21
 
-* TODO: Describe initial release.
+- First Release
+
+## [1.0.3] - 2020-08-13
+
+- Fix Bug $watchZip listen on same type variables
+
+## [1.0.4] - 2020-08-24
+
+- Fix Bug FragmentWidget
+
+## [1.0.5] - 2020-09-30
+## [1.0.6] - 2020-10-05
+
+- Add Constructor to select Broadcast Stream or not